requirements.txt
pyspark==3.1.1
pandas
numpy
pyyaml==3.13
pytest==7.1.2